requirements.txt
numpy==2.1.2 
tqdm==4.67.1 
pyyaml==6.0.2 
matplotlib==3.10.0 
scipy==1.16.1 
scikit-image==0.25.2 
lpips==0.1.4
tabulate==0.9.0